modified:   agent/agent.py 	renamed:    agent/agent_copy.py -> agent/agent_c.py 	modified:   common/arguments.py 	modified:   common/replay_buffer.py 	modified:   common/rollout.py 	modified:   environment/environment.py 	modified:   environment/missile.py 	modified:   main.py 	modified:   requirements.txt

+++ common/replay_buffer.py
@@ -15,7 +15,7 @@ class ReplayBuffer:
         self.current_idx = 0
         self.current_size = 0
         # create the buffer to store info
-        self.buffers = {'o': np.empty([self.size, self.episode_limit, self.n_agents, self.obs_shape]),
+        self.buffers = {'o': np.empty([self.size, self.episode_limit, self.n_agents, self.obs_shape],dtype=np.float32),
                         'u': np.empty([self.size, self.episode_limit, self.n_agents, 1]),
                         's': np.empty([self.size, self.episode_limit, self.state_shape]),
                         'r': np.empty([self.size, self.episode_limit, 1]),
